ExpenseApp/ExpenseApp.Application/Services/GroupService.cs
using ExpenseApp.Core.IRepositories;
using ExpenseApp.Core.Models;
using Microsoft.AspNetCore.Identity;
using Microsoft.EntityFrameworkCore;
using System;
using System.Collections.Generic;
using System.Diagnostics.Metrics;
using System.Linq;
using System.Text;
using System.Threading.Tasks;

namespace ExpenseApp.Application.Services
{
    public class GroupService : IGroupService
    {
        private readonly IGroupRepository _groupRepository;

        public GroupService(IGroupRepository groupRepository)
        {
            _groupRepository = groupRepository;
        }

        public async Task AddMemberToGroupAsync(int groupId, GroupMembers member)
        {
            await _groupRepository.AddMemberToGroupAsync(groupId, member);
        }

        public async Task<Group> CreateGroupAsync(Group group)
        {
            return await _groupRepository.AddGroupAsync(group);
        }

        public async Task DeleteGroupAsync(int groupId)
        {
            await _groupRepository.DeleteGroupAsync(groupId);
        }

        public async Task<IEnumerable<Group>> GetAllGroupsAsync()
        {
            return await _groupRepository.GetAllGroupsAsync();
        }

        public async Task<Group> GetGroupByIdAsync(int groupId)
        {
            return await _groupRepository.GetGroupByIdAsync(groupId);
        }
    }
}
